Inherit the flock file descriptor
