+++ pgctl/cli.py
@@ -24,10 +24,10 @@ from .errors import NoPlayground
 from .errors import PgctlUserMessage
 from .errors import Unsupervised
 from .flock import flock
+from .flock import set_fd_inheritable
 from .functions import commafy
 from .functions import exec_
 from .functions import JSONEncoder
-from .functions import set_fd_inheritable
 from .functions import uniq
 from .service import Service
 from pgctl import __version__
@@ -99,6 +99,7 @@ def pgctl_print(*print_args, **print_kwargs):
     from sys import stderr
     print_kwargs.setdefault('file', stderr)
     print(CHANNEL, *print_args, **print_kwargs)
+    stderr.flush()
 
 
 def timeout(service_name, error, action_name, start_time, timeout_length, check_time):
